requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
asttokens==3.0.0
attrs==25.3.0
awscli==1.42.16
blinker==1.9.0
boto3==1.40.16
botocore==1.40.16
cachetools==5.5.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.2.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.11
decorator==5.1.1
distro==1.9.0
docutils==0.19
executing==2.1.0
filelock==3.19.1
filetype==1.2.0
fonttools==4.55.3
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.7
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mpmath==1.3.0
narwhals==2.1.2
nest-asyncio==1.6.0
networkx==3.5
numpy==2.2.1
openai==1.101.0
opencv-python==4.12.0.88
opencv-python-headless==4.10.0.84
packaging==24.2
pandas==2.2.3
parso==0.8.4
pi_heif==1.1.0
pillow==11.1.0
pillow-avif-plugin==1.5.2
platformdirs==4.3.6
prompt_toolkit==3.0.48
proto-plus==1.26.1
protobuf==5.29.5
psutil==6.1.1
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2024.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.36.2
requests==2.32.3
requests-toolbelt==1.0.0
roboflow==1.2.6
rpds-py==0.27.0
rsa==4.7.2
s3transfer==0.13.1
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
stack-data==0.6.3
streamlit==1.48.1
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
torch==2.8.0
torchvision==0.23.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2024.2
ultralytics==8.3.184
ultralytics-thop==2.0.16
uritemplate==4.2.0
urllib3==2.2.3
watchdog==6.0.0
wcwidth==0.2.13
